move comparison checks to module
